Migrate most components
