--- oldjs/src/components/BetaPage/index.js
@@ -1,4 +1,0 @@
-import BetaInfo from "./BetaInfo";
-import BetaNavBar from "./BetaNavBar";
-
-export { BetaNavBar, BetaInfo };
