--- oldjs/src/components/ChatPage/index.js
@@ -1,4 +1,0 @@
-import Header from './Header';
-import InputBar from './InputBar';
-
-export { Header, InputBar };
