--- oldjs/src/components/ChatPage/mockMessage.jsx
@@ -1,20 +1,0 @@
-export const mockMessages = [
-  {
-    speaker: {
-      name: "Emily",
-      avatarUrl: "https://cdn.discordapp.com/avatars/1106319014919151747/89d6aeae6334cfd7cca06334e4ce323b.webp?size=4096",
-      isHuman: false,
-    },
-    utterance: "Hi! How has your day been?",
-    timestamp: new Date(),
-  },
-  {
-    speaker: {
-      name: "Peepy",
-      avatarUrl: "https://files.catbox.moe/zvv0av.png",
-      isHuman: true,
-    },
-    utterance: "Hey, I'm good.",
-    timestamp: new Date(),
-  },
-];
