--- oldjs/src/components/ChatPage/models/MessageProps.jsx
@@ -1,13 +1,0 @@
-import PropTypes from 'prop-types';
-import Speaker from "./Speaker"; // Make sure this is correct. If Speaker is a custom PropTypes object then import it accordingly.
-
-const Message = ({ speaker, utterance, timestamp }) => {
-}
-
-Message.propTypes = {
-    speaker: PropTypes.instanceOf(Speaker).isRequired,
-    utterance: PropTypes.string.isRequired,
-    timestamp: PropTypes.instanceOf(Date).isRequired
-};
-
-export default Message;
