--- oldjs/src/components/PrivacyPolicyPage/index.js
@@ -1,3 +1,0 @@
-import PrivacySection from "./PrivacySection";
-
-export { PrivacySection };
