--- oldjs/src/components/RegisterPage/index.js
@@ -1,4 +1,0 @@
-import SimpleNavBar from "../Shared/SimpleNavBar";
-import RegisterForm from "./RegisterForm";
-
-export { SimpleNavBar, RegisterForm };
