--- oldjs/src/components/Shared/Contexts/Providers.jsx
@@ -1,17 +1,0 @@
-import { useDarkMode } from './ContextFunctions';
-import { ThemeContext } from './Contexts';
-
-import { Toggle, NavBar } from '../Components';
-
-const Providers = () => {
-  const [isDarkMode, setDarkMode] = useDarkMode('false');
-
-  return (
-    <ThemeContext.Provider value={{ isDarkMode, setDarkMode }}>
-      <Toggle/>
-      <NavBar/>
-    </ThemeContext.Provider>
-  )
-}
-
-export default Providers;
