--- oldjs/src/components/Shared/index.js
@@ -1,6 +1,0 @@
-import TextInput from './TextInput';
-import Button from './Button';
-import Alert from './Alert';
-import NavBar from './NavBar';
-
-export { TextInput, Button, Alert, NavBar };
