--- oldjs/src/hoc/SectionWrapper.jsx
@@ -1,23 +1,0 @@
-import { motion } from "framer-motion";
-
-import { staggerContainer } from "../utils/motion";
-
-const SectionWrapper = (Component, idName) =>
-  function HOC() {
-    return (
-      <motion.section
-        variants={staggerContainer()}
-        initial="hidden"
-        whileInView="show"
-        viewport={{ once: true, amount: 0.25 }}
-        className={`p-custom max-w-7xl, relative z-0 mx-auto`}
-      >
-        <span className="mt-[-100px] block pb-[100px]" id={idName}>
-          &nbsp;
-        </span>
-        <Component />
-      </motion.section>
-    );
-  };
-
-export default SectionWrapper;
