+++ src/components/BetaPage/index.ts
@@ -1,0 +1,4 @@
+import { BetaInfo } from "./BetaInfo";
+import {BetaNavBar} from "./BetaNavBar";
+
+export { BetaInfo, BetaNavBar };
