+++ src/components/PrivacyPolicyPage/index.ts
@@ -1,0 +1,3 @@
+import {PrivacySection} from "./PrivacySection";
+
+export { PrivacySection };
