+++ src/components/RegisterPage/index.ts
@@ -1,0 +1,4 @@
+import { SimpleNavBar } from "../Shared/SimpleNavBar";
+import { RegisterForm } from "./RegisterForm";
+
+export { RegisterForm, SimpleNavBar };
