+++ src/components/Shared/Contexts/Providers.tsx
@@ -1,0 +1,7 @@
+import { NavBar } from "../Components";
+
+const Providers = () => {
+	return <NavBar />;
+};
+
+export default Providers;
