+++ src/components/Shared/index.ts
@@ -1,0 +1,6 @@
+import { Alert } from "./Alert";
+import { Button } from "./Button";
+import { NavBar } from "./NavBar";
+import { TextInput } from "./TextInput";
+
+export { Alert, Button, NavBar, TextInput };
