+++ src/components/TermsOfServicePage/index.ts
